% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAreaAndCalculateMapAccuracy.R
\name{estimateAreaAndCalculateMapAccuracy}
\alias{estimateAreaAndCalculateMapAccuracy}
\title{estimateAreaAndCalculateMapAccuracy function}
\usage{
estimateAreaAndCalculateMapAccuracy(
  counts = system.file("counts", "counts.csv", package = "3I3D-GEE_areaEstimator"),
  sample = system.file("sample", "sample.shp", package = "3I3D-GEE_areaEstimator"),
  outPath = getwd()
)
}
\value{
list
}
\description{
Calculate area estimates and map accuracy using 3I3D-GEE outputs
}
\examples{
estimateAreaAndCalculateMapAccuracy()
}
\keyword{mapAccuracy}
